commit 13 : add account fragment, other UI

+++ app/src/main/java/com/example/grocerystore/ui/activityMain/fragments/firstTab/adapters/TitleUIStateAdapter.kt
@@ -32,7 +32,7 @@ class TitleUIStateAdapter(list: List<TitleUIState>, private val context: Context
 
         override fun onBindViewHolder(holder: ItemViewHolder, position: Int) = holder.bind(data[position])
 
-        override fun getItemId(position: Int) = data[position].id
+        override fun getItemId(position: Int) = data[position].id.toLong()
 
         override fun getItemCount() = data.size
